# ecommerce-app/openapi.yaml
openapi: 3.1.0
...
info:
  title: ShopKart E-Commerce Storefront API
  version: 1.0.0
  summary: Customer-facing storefront API for the ShopKart e-commerce application.
  description: |
    REST API powering the ShopKart storefront: authentication, profile & address
    management, product catalogue browsing, cart & coupon handling, checkout /
    orders, and verified product reviews.

    ## Authentication
    Stateless **JWT Bearer** tokens. Obtain a token from `POST /api/auth/signup`
    or `POST /api/auth/login`, then send it as `Authorization: Bearer <token>`.
    Tokens expire after 7 days (configurable via `JWT_EXPIRES_IN`). Logout is
    client-side (drop the token); `POST /api/auth/logout` is a no-op convenience.

    ## Core domain rule
    A purchase is recorded immutably as `user -> order -> order_items`. Each
    `order_item` snapshots the product name, variant name, SKU and unit price
    paid at that time, so order history never changes if a product is later
    renamed, repriced or archived.

    ## Pricing
    - Tax: 8% of `(subtotal - discount)`.
    - Shipping: flat ₹49, waived when `subtotal >= ₹999` (or cart is empty).
    - Coupons: `percent` (optionally capped via `max_discount`) or `fixed`.

    ## Conventions
    - All request and response bodies are `application/json` unless noted.
    - All monetary values are in INR (₹).
    - Timestamps are RFC 3339 / ISO 8601 date-time in UTC.
  contact:
    name: ShopKart API Support
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/products/{productId}/reviews:
    get:
      tags: [Reviews]
      summary: List reviews for a product
      description: Public list of a product's reviews, newest first.
      operationId: listProductReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: The product's reviews.
          content:
            application/json:
              schema:
                type: object
                required: [reviews]
                properties:
                  reviews:
                    type: array
                    items: { $ref: '#/components/schemas/Review' }
              example:
                reviews:
                  - id: 5
                    rating: 5
                    title: Fantastic sound
                    body: Battery really does last all week.
                    created_at: '2026-07-20T14:12:00Z'
                    updated_at: '2026-07-20T14:12:00Z'
                    author: Aarav Sharma

  # ───────────────────────────────  CART  ───────────────────────────────
...
  /api/reviews/eligibility/{productId}:
    get:
      tags: [Reviews]
      summary: Check review eligibility for a product
      description: |
        Tells the client whether the authenticated user may review a product
        (they must have received it in a `delivered` order and not already
        reviewed it), plus any existing review and whether it is still editable.
      operationId: getReviewEligibility
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Eligibility details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewEligibility' }
              examples:
                canReview:
                  value: { can_review: true, has_delivered: true, review: null, can_edit: false }
                alreadyReviewed:
                  value:
                    can_review: false
                    has_delivered: true
                    can_edit: true
                    review:
                      id: 5
                      product_id: 1
                      user_id: 42
                      order_item_id: 900
                      rating: 5
                      title: Fantastic sound
                      body: Battery really does last all week.
                      created_at: '2026-07-23T12:00:00Z'
                      updated_at: '2026-07-23T12:00:00Z'
                      edit_deadline: '2026-07-23T12:30:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reviews:
    post:
      tags: [Reviews]
      summary: Write a review for a delivered product
      description: |
        Creates a review. The user must have received the product in a
        `delivered` order and must not have reviewed it before (one review per
        product per user). Reviews are editable for 30 minutes after posting.
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewCreateRequest' }
            example: { product_id: 1, rating: 5, title: Fantastic sound, body: Battery lasts all week. }
      responses:
        '201':
          description: Review created.
          content:
            application/json:
              schema:
                type: object
                required: [review]
                properties:
                  review: { $ref: '#/components/schemas/ReviewFull' }
        '400':
          description: Missing product or rating out of range.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                product: { value: { error: 'Product is required.' } }
                rating: { value: { error: 'Rating must be between 1 and 5 stars.' } }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The user has not received this product in a delivered order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'You can only review a product you have received in a past order.' }
        '409':
          description: The user has already reviewed this product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'You have already reviewed this product.' }
  /api/reviews/{id}:
    parameters:
      - $ref: '#/components/parameters/ReviewId'
    put:
      tags: [Reviews]
      summary: Edit my review (within the 30-minute window)
      description: |
        Updates the caller's own review. Only allowed before `edit_deadline`
        (30 minutes after creation). Any omitted field keeps its current value.
      operationId: updateReview
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewUpdateRequest' }
            example: { rating: 4, title: Great, body: Downgraded half a star for the case. }
      responses:
        '200':
          description: Review updated.
          content:
            application/json:
              schema:
                type: object
                required: [review]
                properties:
                  review: { $ref: '#/components/schemas/ReviewFull' }
        '400':
          description: Rating out of range.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'Rating must be between 1 and 5 stars.' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The review belongs to another user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'You can only edit your own review.' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The edit window has closed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'The edit window for this review has closed.' }

# ============================================================================
#  COMPONENTS
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from `/api/auth/login` or `/api/auth/signup`.
        Send as `Authorization: Bearer <token>`. Payload carries `id`, `email`,
        `role` and `name`; expires after 7 days by default.

  # ─────────────────────────────  PARAMETERS  ─────────────────────────────
  parameters:
...
    ProductId:
      name: productId
      in: path
      required: true
      description: Numeric product identifier.
      schema: { type: integer, format: int32, minimum: 1 }
      example: 1
    ReviewId:
      name: id
      in: path
      required: true
      description: Review identifier.
      schema: { type: integer, format: int32, minimum: 1 }
      example: 5
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the token is invalid/expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missing: { value: { error: 'Please log in to continue.' } }
            expired: { value: { error: 'Your session has expired. Please log in again.' } }
    NotFound:
      description: The requested resource does not exist (or is not owned by the caller).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Not found.' }
...
  schemas:
    Error:
      type: object
      description: |
        Standard error body returned by every failing endpoint. The API returns
        a single human-readable `error` message (no code/details envelope).
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Human-readable, client-safe error message.
          example: 'Please log in to continue.'
...
    Review:
      type: object
      description: Public-facing review (author name only).
      required: [id, rating, created_at, author]
      properties:
        id: { type: integer, format: int32, example: 5 }
        rating: { type: integer, minimum: 1, maximum: 5, example: 5 }
        title: { type: [string, 'null'], example: Fantastic sound }
        body: { type: [string, 'null'], example: Battery really does last all week. }
        created_at: { type: string, format: date-time, example: '2026-07-20T14:12:00Z' }
        updated_at: { type: string, format: date-time, example: '2026-07-20T14:12:00Z' }
        author: { type: string, description: Display name of the reviewer., example: Aarav Sharma }
    ReviewFull:
      type: object
      description: Full review row returned to the owner on create/edit.
      required: [id, product_id, user_id, rating, created_at, updated_at, edit_deadline]
      properties:
        id: { type: integer, format: int32, example: 5 }
        product_id: { type: integer, format: int32, example: 1 }
        user_id: { type: integer, format: int32, example: 42 }
        order_item_id: { type: [integer, 'null'], format: int32, description: The delivered order item that entitles this review., example: 900 }
        rating: { type: integer, minimum: 1, maximum: 5, example: 5 }
        title: { type: [string, 'null'], example: Fantastic sound }
        body: { type: [string, 'null'], example: Battery really does last all week. }
        created_at: { type: string, format: date-time, example: '2026-07-23T12:00:00Z' }
        updated_at: { type: string, format: date-time, example: '2026-07-23T12:00:00Z' }
        edit_deadline:
          type: string
          format: date-time
          description: After this instant the review can no longer be edited (30 min after creation).
          example: '2026-07-23T12:30:00Z'
...
    ReviewEligibility:
      type: object
      required: [can_review, has_delivered, review, can_edit]
      properties:
        can_review:
          type: boolean
          description: True when the user has a delivered purchase and no existing review.
          example: true
        has_delivered:
          type: boolean
          description: True when the user has received this product in a delivered order.
          example: true
        review:
          oneOf:
            - { $ref: '#/components/schemas/ReviewFull' }
            - { type: 'null' }
          description: The user's existing review, or null.
        can_edit:
          type: boolean
          description: True when an existing review is still within its edit window.
          example: false
    ReviewCreateRequest:
      type: object
      required: [product_id, rating]
      additionalProperties: false
      properties:
        product_id: { type: integer, format: int32, minimum: 1, example: 1 }
        rating: { type: integer, minimum: 1, maximum: 5, example: 5 }
        title: { type: [string, 'null'], example: Fantastic sound }
        body: { type: [string, 'null'], example: Battery really does last all week. }
    ReviewUpdateRequest:
      type: object
      description: All fields optional; omitted fields keep their current value.
      additionalProperties: false
      properties:
        rating: { type: integer, minimum: 1, maximum: 5, example: 4 }
        title: { type: [string, 'null'], example: Great }
        body: { type: [string, 'null'], example: Downgraded half a star for the case. }